fix all problems identified by clang static analyzer

+++ src/eth_plugin_interface.h
@@ -134,9 +134,9 @@ typedef struct ethQueryContractID_t {
     uint8_t *pluginContext;
 
     char *name;
-    uint32_t nameLength;
+    size_t nameLength;
     char *version;
-    uint32_t versionLength;
+    size_t versionLength;
 
     uint8_t result;
 
@@ -150,9 +150,9 @@ typedef struct ethQueryContractUI_t {
     uint8_t *pluginContext;
     uint8_t screenIndex;
     char *title;
-    uint32_t titleLength;
+    size_t titleLength;
     char *msg;
-    uint32_t msgLength;
+    size_t msgLength;
 
     uint8_t result;
